tzdb-package {tzdb}	R Documentation

tzdb: Time Zone Database Information

Description

Provides an up-to-date copy of the Internet Assigned Numbers Authority (IANA) Time Zone Database. It is updated periodically to reflect changes made by political bodies to time zone boundaries, UTC offsets, and daylight saving time rules. Additionally, this package provides a C++ interface for working with the 'date' library. 'date' provides comprehensive support for working with dates and date-times, which this package exposes to make it easier for other R packages to utilize. Headers are provided for calendar specific calculations, along with a limited interface for time zone manipulations.

Author(s)

Maintainer: Davis Vaughan davis@rstudio.com

Other contributors:

Howard Hinnant (Author of the included date library) [copyright holder]

RStudio [copyright holder, funder]

See Also

Useful links:

https://r-lib.github.io/tzdb/

https://github.com/r-lib/tzdb

Report bugs at https://github.com/r-lib/tzdb/issues

[Package tzdb version 0.2.0 Index]