tzdb_path {tzdb}	R Documentation

Retrieve the path to the time zone database

Description

Retrieve the path to the time zone database

Usage

tzdb_path(type)

Arguments

type
[character(1)]

The type of time zone database to return the path for.

Currently only "text" is supported.

Value

A single string containing the path to the database.

Examples

tzdb_path("text")

[Package tzdb version 0.2.0 Index]